+++ protected/config/main.php
@@ -35,6 +35,7 @@ return [
     'sourceLanguage' => 'en',
     // Default theme
     'theme' => 'default',
+    'layout' => 'yupe',
     'charset' => 'UTF-8',
     'controllerNamespace' => 'application\controllers',
     'preload' => defined('YII_DEBUG') && YII_DEBUG ? ['debug'] : [],