+++ themes/default/views/layouts/column1.php
@@ -1,8 +1,0 @@
-<?php $this->beginContent('//layouts/main'); ?>
-<div class="container">
-    <div id="content">
-        <?= $content; ?>
-    </div>
-    <!-- content -->
-</div>
-<?php $this->endContent(); ?>